requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
requests>=2.28.0
aiohttp>=3.8.0
tqdm>=4.64.0
jinja2>=3.1.0

# Type checking support
typing-extensions>=4.0.0

# Data validation
pydantic>=2.0.0

# File handling
pathlib2>=2.3.0; python_version < "3.10"

# Checksum and crypto
hashlib>=3.6; python_version >= "3.6"

# Process management
psutil>=5.9.0

# Async support
asyncio>=3.4; python_version < "3.7"
